refactor(codec): own sequence ordering runtime
