+++ crates/copybook-codec/Cargo.toml
@@ -37,7 +37,6 @@ tracing.workspace = true
 base64.workspace = true
 sha2.workspace = true
 blake3.workspace = true
-copybook-sequence-ring.workspace = true
 crossbeam-channel.workspace = true
 smallvec.workspace = true
 indexmap.workspace = true
