+++ crates/copybook-sequence-ring/Cargo.toml
@@ -20,8 +20,10 @@ all-features = false
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
+copybook-codec.workspace = true
+
+[dev-dependencies]
 crossbeam-channel.workspace = true
-tracing.workspace = true
 
 [lints]
 workspace = true
