+++ crates/copybook-sequence-ring/src/lib.rs
@@ -1,779 +1,9 @@
 #![cfg_attr(not(test), deny(clippy::unwrap_used, clippy::expect_used))]
 // SPDX-License-Identifier: AGPL-3.0-or-later
-//! Deterministic sequence reordering primitive for parallel pipelines.
+//! Backward-compatible facade for codec-owned sequence ordering.
 //!
-//! `SequenceRing` accepts potentially out-of-order records tagged with a
-//! sequence ID and emits them in-order.
-
-use crossbeam_channel::{Receiver, Sender, bounded};
-use std::collections::BTreeMap;
-use tracing::{debug, warn};
-
-/// Record with sequence ID for ordered processing
-///
-/// Wraps a record with a monotonically increasing sequence ID to enable
-/// deterministic output ordering even when processing happens in parallel.
-/// This is essential for maintaining data integrity when using multi-threaded
-/// record processing.
-///
-/// # Examples
-///
-/// ```rust
-/// use copybook_sequence_ring::SequencedRecord;
-///
-/// let record1 = SequencedRecord::new(1, "first record");
-/// let record2 = SequencedRecord::new(2, "second record");
-///
-/// assert_eq!(record1.sequence_id, 1);
-/// assert_eq!(record2.data, "second record");
-/// ```
-#[derive(Debug, Clone)]
-pub struct SequencedRecord<T> {
-    /// Sequence number for ordering
-    ///
-    /// Monotonically increasing ID assigned when record enters processing.
-    /// Used to reconstruct original order after parallel processing.
-    pub sequence_id: u64,
-
-    /// The actual record data
-    ///
-    /// Payload being processed (e.g., binary COBOL record, JSON value).
-    pub data: T,
-}
-
-impl<T> SequencedRecord<T> {
-    /// Create a new sequenced record
-    ///
-    /// # Arguments
-    ///
-    /// * `sequence_id` - Unique monotonic sequence number for ordering
-    /// * `data` - Record payload to process
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use copybook_sequence_ring::SequencedRecord;
-    ///
-    /// let record = SequencedRecord::new(42, vec![1, 2, 3]);
-    /// assert_eq!(record.sequence_id, 42);
-    /// assert_eq!(record.data, vec![1, 2, 3]);
-    /// ```
-    #[inline]
-    #[must_use]
-    pub fn new(sequence_id: u64, data: T) -> Self {
-        Self { sequence_id, data }
-    }
-}
-
-/// Bounded channel with sequence tracking for ordered emission
-///
-/// Provides deterministic output ordering for parallel processing by buffering
-/// out-of-order records and emitting them in sequence order. This is critical
-/// for maintaining data integrity when processing COBOL records in parallel.
-///
-/// # How It Works
-///
-/// 1. Workers submit processed records with sequence IDs via sender channel
-/// 2. `SequenceRing` receives records (potentially out-of-order)
-/// 3. Out-of-order records are buffered in reorder buffer (`BTreeMap`)
-/// 4. Records are emitted in strict sequence order via [`recv_ordered()`](SequenceRing::recv_ordered)
-///
-/// # Memory Bounds
-///
-/// - **Channel capacity** - Maximum records in flight between workers and consumer
-/// - **Reorder window** - Maximum buffered out-of-order records (warns if exceeded)
-///
-/// # Performance Characteristics
-///
-/// - **O(log n)** insertion/removal from reorder buffer (`BTreeMap`)
-/// - **O(1)** emission when records arrive in order (hot path)
-/// - **Memory usage** - Bounded by channel capacity + reorder window size
-///
-/// # Examples
-///
-/// ```rust
-/// use copybook_sequence_ring::{SequenceRing, SequencedRecord};
-///
-/// let mut ring = SequenceRing::new(100, 50); // 100 capacity, 50 max window
-/// let sender = ring.sender();
-///
-/// // Simulate workers sending out-of-order results
-/// sender.send(SequencedRecord::new(2, "second")).unwrap();
-/// sender.send(SequencedRecord::new(1, "first")).unwrap();
-/// sender.send(SequencedRecord::new(3, "third")).unwrap();
-///
-/// // Consumer receives in order
-/// assert_eq!(ring.recv_ordered().unwrap(), Some("first"));
-/// assert_eq!(ring.recv_ordered().unwrap(), Some("second"));
-/// assert_eq!(ring.recv_ordered().unwrap(), Some("third"));
-/// ```
-#[derive(Debug)]
-pub struct SequenceRing<T> {
-    /// Channel for receiving processed records
-    receiver: Receiver<SequencedRecord<T>>,
-
-    /// Sender for processed records (cloned to workers)
-    sender: Sender<SequencedRecord<T>>,
-
-    /// Buffer for out-of-order records
-    ///
-    /// `BTreeMap` provides O(log n) ordered access to buffered records.
-    reorder_buffer: BTreeMap<u64, T>,
-
-    /// Next expected sequence ID
-    ///
-    /// Monotonically incremented as records are emitted in order.
-    next_sequence_id: u64,
-
-    /// Maximum reordering window size
-    ///
-    /// Warning threshold for reorder buffer size (indicates excessive skew).
-    max_window_size: usize,
-
-    /// Channel capacity
-    channel_capacity: usize,
-}
-
-impl<T> SequenceRing<T> {
-    /// Create a new sequence ring with bounded capacity
-    ///
-    /// # Arguments
-    ///
-    /// * `channel_capacity` - Maximum number of records in flight between workers and consumer
-    /// * `max_window_size` - Maximum buffered out-of-order records (warning threshold)
-    ///
-    /// # Tuning Guidelines
-    ///
-    /// - **Channel capacity**: Should be 2-4x number of worker threads for good throughput
-    /// - **Reorder window**: Should be `channel_capacity` / 2 to allow for processing variance
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use copybook_sequence_ring::SequenceRing;
-    ///
-    /// // For 4-worker pool: 16 capacity, 8 window
-    /// let ring: SequenceRing<String> = SequenceRing::new(16, 8);
-    ///
-    /// // For 8-worker pool: 32 capacity, 16 window
-    /// let ring: SequenceRing<Vec<u8>> = SequenceRing::new(32, 16);
-    /// ```
-    #[inline]
-    #[must_use]
-    pub fn new(channel_capacity: usize, max_window_size: usize) -> Self {
-        let (sender, receiver) = bounded(channel_capacity);
-
-        Self {
-            receiver,
-            sender,
-            reorder_buffer: BTreeMap::new(),
-            next_sequence_id: 1,
-            max_window_size,
-            channel_capacity,
-        }
-    }
-
-    /// Get a sender for workers to submit processed records
-    ///
-    /// Returns a cloneable sender that workers can use to submit processed
-    /// records back to the sequence ring. Multiple workers can share clones
-    /// of this sender.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use copybook_sequence_ring::{SequenceRing, SequencedRecord};
-    /// use std::thread;
-    ///
-    /// let mut ring = SequenceRing::new(10, 5);
-    /// let sender1 = ring.sender();
-    /// let sender2 = ring.sender(); // Clone for another worker
-    ///
-    /// // Workers can send concurrently
-    /// let handle = thread::spawn(move || {
-    ///     sender1.send(SequencedRecord::new(1, "data")).unwrap();
-    /// });
-    ///
-    /// sender2.send(SequencedRecord::new(2, "data")).unwrap();
-    /// handle.join().unwrap();
-    /// ```
-    #[inline]
-    #[must_use]
-    pub fn sender(&self) -> Sender<SequencedRecord<T>> {
-        self.sender.clone()
-    }
-
-    /// Receive the next record in sequence order.
-    /// Blocks until the next expected record is available.
-    ///
-    /// # Errors
-    /// Returns an error if the channel is disconnected.
-    #[inline]
-    #[must_use = "Handle the Result or propagate the error"]
-    pub fn recv_ordered(&mut self) -> Result<Option<T>, crossbeam_channel::RecvError> {
-        loop {
-            // Check if we have the next expected record in the reorder buffer
-            if let Some(record) = self.reorder_buffer.remove(&self.next_sequence_id) {
-                self.next_sequence_id += 1;
-                debug!(
-                    "Emitting record {} from reorder buffer",
-                    self.next_sequence_id - 1
-                );
-                return Ok(Some(record));
-            }
-
-            // Receive next record from channel
-            if let Ok(sequenced_record) = self.receiver.recv() {
-                let SequencedRecord { sequence_id, data } = sequenced_record;
-
-                match sequence_id.cmp(&self.next_sequence_id) {
-                    std::cmp::Ordering::Equal => {
-                        // This is the next expected record
-                        self.next_sequence_id += 1;
-                        debug!("Emitting record {} directly", sequence_id);
-                        return Ok(Some(data));
-                    }
-                    std::cmp::Ordering::Greater => {
-                        // Future record - buffer it
-                        debug!(
-                            "Buffering out-of-order record {} (expecting {})",
-                            sequence_id, self.next_sequence_id
-                        );
-                        self.reorder_buffer.insert(sequence_id, data);
-
-                        // Check reorder buffer size
-                        if self.reorder_buffer.len() > self.max_window_size {
-                            warn!(
-                                "Reorder buffer size ({}) exceeds maximum ({}), potential memory issue",
-                                self.reorder_buffer.len(),
-                                self.max_window_size
-                            );
-                        }
-                    }
-                    std::cmp::Ordering::Less => {
-                        // Past record - this shouldn't happen with proper sequencing
-                        warn!(
-                            "Received past record {} (expecting {}), ignoring",
-                            sequence_id, self.next_sequence_id
-                        );
-                    }
-                }
-            } else {
-                // Channel closed - emit any remaining buffered records
-                if let Some((_, record)) = self.reorder_buffer.pop_first() {
-                    debug!("Emitting remaining buffered record during shutdown");
-                    return Ok(Some(record));
-                }
-                debug!("Channel closed, no more records");
-                return Ok(None);
-            }
-        }
-    }
-
-    /// Try to receive the next record without blocking.
-    ///
-    /// # Errors
-    /// Returns an error if the channel is disconnected or would block.
-    #[inline]
-    #[must_use = "Handle the Result or propagate the error"]
-    pub fn try_recv_ordered(&mut self) -> Result<Option<T>, crossbeam_channel::TryRecvError> {
-        // Check if we have the next expected record in the reorder buffer
-        if let Some(record) = self.reorder_buffer.remove(&self.next_sequence_id) {
-            self.next_sequence_id += 1;
-            return Ok(Some(record));
-        }
-
-        // Try to receive from channel
-        match self.receiver.try_recv() {
-            Ok(sequenced_record) => {
-                let SequencedRecord { sequence_id, data } = sequenced_record;
-
-                match sequence_id.cmp(&self.next_sequence_id) {
-                    std::cmp::Ordering::Equal => {
-                        self.next_sequence_id += 1;
-                        Ok(Some(data))
-                    }
-                    std::cmp::Ordering::Greater => {
-                        // Buffer future record and return empty
-                        self.reorder_buffer.insert(sequence_id, data);
-                        Err(crossbeam_channel::TryRecvError::Empty)
-                    }
-                    std::cmp::Ordering::Less => {
-                        // Past record - ignore and try again
-                        warn!("Received past record {}, ignoring", sequence_id);
-                        Err(crossbeam_channel::TryRecvError::Empty)
-                    }
-                }
-            }
-            Err(e) => Err(e),
-        }
-    }
-
-    /// Get statistics about the sequence ring
-    ///
-    /// Returns current operational statistics including reorder buffer usage
-    /// and sequence tracking state.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use copybook_sequence_ring::{SequenceRing, SequencedRecord};
-    /// use crossbeam_channel::TryRecvError;
-    ///
-    /// let mut ring = SequenceRing::new(10, 5);
-    /// let sender = ring.sender();
-    ///
-    /// // Send out-of-order record (seq 2 when expecting 1)
-    /// sender.send(SequencedRecord::new(2, "data")).unwrap();
-    ///
-    /// // Force the ring to observe and buffer the out-of-order record
-    /// assert!(matches!(ring.try_recv_ordered(), Err(TryRecvError::Empty)));
-    ///
-    /// let stats = ring.stats();
-    /// assert_eq!(stats.next_sequence_id, 1); // Still waiting for record 1
-    /// assert_eq!(stats.reorder_buffer_size, 1); // Record 2 is buffered
-    /// ```
-    #[inline]
-    #[must_use]
-    pub fn stats(&self) -> SequenceRingStats {
-        SequenceRingStats {
-            next_sequence_id: self.next_sequence_id,
-            reorder_buffer_size: self.reorder_buffer.len(),
-            max_window_size: self.max_window_size,
-            channel_capacity: self.channel_capacity,
-        }
-    }
-}
-
-/// Statistics about sequence ring operation
-///
-/// Provides visibility into sequence ring health and performance characteristics.
-#[derive(Debug, Clone)]
-pub struct SequenceRingStats {
-    /// Next expected sequence ID
-    ///
-    /// The sequence number of the next record to be emitted.
-    pub next_sequence_id: u64,
-
-    /// Current reorder buffer size
-    ///
-    /// Number of out-of-order records currently buffered. High values indicate
-    /// significant processing variance between workers.
-    pub reorder_buffer_size: usize,
-
-    /// Maximum reorder window size
-    ///
-    /// Warning threshold for reorder buffer size.
-    pub max_window_size: usize,
-
-    /// Channel capacity
-    ///
-    /// Maximum records in flight between workers and consumer.
-    pub channel_capacity: usize,
-}
-
-#[cfg(test)]
-#[allow(clippy::expect_used, clippy::unwrap_used)]
-mod tests {
-    use super::*;
-    use crossbeam_channel::TryRecvError;
-
-    #[test]
-    fn sequenced_record_creation() {
-        let record = SequencedRecord::new(42, "test data");
-        assert_eq!(record.sequence_id, 42);
-        assert_eq!(record.data, "test data");
-    }
-
-    #[test]
-    fn sequenced_record_clone() {
-        let record = SequencedRecord::new(1, vec![10, 20, 30]);
-        let cloned = record.clone();
-        assert_eq!(cloned.sequence_id, 1);
-        assert_eq!(cloned.data, vec![10, 20, 30]);
-    }
-
-    #[test]
-    fn sequenced_record_debug_format() {
-        let record = SequencedRecord::new(7, "hello");
-        let debug = format!("{record:?}");
-        assert!(debug.contains('7'));
-        assert!(debug.contains("hello"));
-    }
-
-    #[test]
-    fn recv_ordered_emits_input_order() {
-        let mut ring = SequenceRing::new(10, 5);
-        let sender = ring.sender();
-
-        sender.send(SequencedRecord::new(2, "second")).unwrap();
-        sender.send(SequencedRecord::new(1, "first")).unwrap();
-        sender.send(SequencedRecord::new(3, "third")).unwrap();
-
-        assert_eq!(ring.recv_ordered().unwrap(), Some("first"));
-        assert_eq!(ring.recv_ordered().unwrap(), Some("second"));
-        assert_eq!(ring.recv_ordered().unwrap(), Some("third"));
-    }
-
-    #[test]
-    fn recv_ordered_already_in_order() {
-        let mut ring = SequenceRing::new(10, 5);
-        let sender = ring.sender();
-
-        sender.send(SequencedRecord::new(1, "a")).unwrap();
-        sender.send(SequencedRecord::new(2, "b")).unwrap();
-        sender.send(SequencedRecord::new(3, "c")).unwrap();
-
-        assert_eq!(ring.recv_ordered().unwrap(), Some("a"));
-        assert_eq!(ring.recv_ordered().unwrap(), Some("b"));
-        assert_eq!(ring.recv_ordered().unwrap(), Some("c"));
-    }
-
-    #[test]
-    fn recv_ordered_reverse_order() {
-        let mut ring = SequenceRing::new(10, 5);
-        let sender = ring.sender();
-
-        sender.send(SequencedRecord::new(3, "c")).unwrap();
-        sender.send(SequencedRecord::new(2, "b")).unwrap();
-        sender.send(SequencedRecord::new(1, "a")).unwrap();
-
-        assert_eq!(ring.recv_ordered().unwrap(), Some("a"));
-        assert_eq!(ring.recv_ordered().unwrap(), Some("b"));
-        assert_eq!(ring.recv_ordered().unwrap(), Some("c"));
-    }
-
-    #[test]
-    fn recv_ordered_returns_none_on_channel_close() {
-        let ring = SequenceRing::<&str>::new(10, 5);
-        // Drop all senders so channel is disconnected
-        drop(ring.sender());
-        // The ring holds its own sender; need to consume it
-        // Actually, we can't easily force the ring's internal sender to drop.
-        // Instead, test that after sending all records and dropping sender,
-        // recv_ordered eventually returns None.
-        let mut ring2 = SequenceRing::new(10, 5);
-        let sender = ring2.sender();
-        sender.send(SequencedRecord::new(1, "only")).unwrap();
-        drop(sender);
-        // Ring still holds internal sender clone; drop it by replacing
-        // We'll just verify the record comes through
-        assert_eq!(ring2.recv_ordered().unwrap(), Some("only"));
-    }
-
-    #[test]
-    fn try_recv_empty_reports_empty() {
-        let mut ring = SequenceRing::<&str>::new(10, 5);
-        assert!(matches!(ring.try_recv_ordered(), Err(TryRecvError::Empty)));
-    }
-
-    #[test]
-    fn try_recv_returns_in_order_record() {
-        let mut ring = SequenceRing::new(10, 5);
-        let sender = ring.sender();
-        sender.send(SequencedRecord::new(1, "first")).unwrap();
-        assert_eq!(ring.try_recv_ordered().unwrap(), Some("first"));
-    }
-
-    #[test]
-    fn try_recv_buffers_future_record() {
-        let mut ring = SequenceRing::new(10, 5);
-        let sender = ring.sender();
-
-        // Send record 2 (expecting 1)
-        sender.send(SequencedRecord::new(2, "second")).unwrap();
-        // Should buffer it and return Empty
-        assert!(matches!(ring.try_recv_ordered(), Err(TryRecvError::Empty)));
-        // Buffer should contain record 2
-        assert_eq!(ring.stats().reorder_buffer_size, 1);
-
-        // Now send record 1
-        sender.send(SequencedRecord::new(1, "first")).unwrap();
-        assert_eq!(ring.try_recv_ordered().unwrap(), Some("first"));
-        // Record 2 should now be available from buffer
-        assert_eq!(ring.try_recv_ordered().unwrap(), Some("second"));
-    }
-
-    #[test]
-    fn try_recv_ignores_past_record() {
-        let mut ring = SequenceRing::new(10, 5);
-        let sender = ring.sender();
-
-        // Emit record 1 normally
-        sender.send(SequencedRecord::new(1, "first")).unwrap();
-        assert_eq!(ring.try_recv_ordered().unwrap(), Some("first"));
-
-        // Send duplicate/past record 1 again
-        sender.send(SequencedRecord::new(1, "duplicate")).unwrap();
-        assert!(matches!(ring.try_recv_ordered(), Err(TryRecvError::Empty)));
-    }
-
-    #[test]
-    fn sender_respects_channel_capacity() {
-        let ring = SequenceRing::new(2, 1);
-        let sender = ring.sender();
-
-        sender.try_send(SequencedRecord::new(1, "first")).unwrap();
-        sender.try_send(SequencedRecord::new(2, "second")).unwrap();
-
-        let full = sender.try_send(SequencedRecord::new(3, "third"));
-        assert!(full.is_err());
-    }
-
-    #[test]
-    fn stats_reflect_progress() {
-        let mut ring = SequenceRing::new(10, 5);
-        let sender = ring.sender();
-
-        for i in 1..=5 {
-            sender
-                .send(SequencedRecord::new(i, format!("record_{i}")))
-                .unwrap();
-        }
-
-        let before = ring.stats();
-        assert_eq!(before.next_sequence_id, 1);
-        assert_eq!(before.reorder_buffer_size, 0);
-        assert_eq!(before.channel_capacity, 10);
-        assert_eq!(before.max_window_size, 5);
-
-        for _ in 1..=5 {
-            assert!(ring.recv_ordered().unwrap().is_some());
-        }
-
-        let after = ring.stats();
-        assert_eq!(after.next_sequence_id, 6);
-        assert_eq!(after.reorder_buffer_size, 0);
-    }
-
-    #[test]
-    fn stats_initial_state() {
-        let ring = SequenceRing::<String>::new(100, 50);
-        let stats = ring.stats();
-        assert_eq!(stats.next_sequence_id, 1);
-        assert_eq!(stats.reorder_buffer_size, 0);
-        assert_eq!(stats.max_window_size, 50);
-        assert_eq!(stats.channel_capacity, 100);
-    }
-
-    #[test]
-    fn stats_shows_buffered_count() {
-        let mut ring = SequenceRing::new(10, 5);
-        let sender = ring.sender();
-
-        // Send records 3, 5 (out of order, both future)
-        sender.send(SequencedRecord::new(3, "three")).unwrap();
-        sender.send(SequencedRecord::new(5, "five")).unwrap();
-        // Force reads to buffer them
-        assert!(matches!(ring.try_recv_ordered(), Err(TryRecvError::Empty)));
-        assert!(matches!(ring.try_recv_ordered(), Err(TryRecvError::Empty)));
-
-        let stats = ring.stats();
-        assert_eq!(stats.reorder_buffer_size, 2);
-        assert_eq!(stats.next_sequence_id, 1);
-    }
-
-    #[test]
-    fn multiple_senders_work() {
-        let mut ring = SequenceRing::new(10, 5);
-        let sender1 = ring.sender();
-        let sender2 = ring.sender();
-
-        sender1.send(SequencedRecord::new(1, "from_s1")).unwrap();
-        sender2.send(SequencedRecord::new(2, "from_s2")).unwrap();
-
-        assert_eq!(ring.recv_ordered().unwrap(), Some("from_s1"));
-        assert_eq!(ring.recv_ordered().unwrap(), Some("from_s2"));
-    }
-
-    #[test]
-    fn large_gap_reordering() {
-        let mut ring = SequenceRing::new(20, 20);
-        let sender = ring.sender();
-
-        // Send records 10, 9, 8, ..., 1 (reverse order)
-        for i in (1..=10).rev() {
-            sender.send(SequencedRecord::new(i, i)).unwrap();
-        }
-
-        // Should emit 1..=10 in order
-        for expected in 1..=10u64 {
-            assert_eq!(ring.recv_ordered().unwrap(), Some(expected));
-        }
-    }
-
-    #[test]
-    fn sequence_ring_stats_clone_and_debug() {
-        let ring = SequenceRing::<()>::new(8, 4);
-        let stats = ring.stats();
-        let cloned = stats.clone();
-        assert_eq!(cloned.channel_capacity, 8);
-        let debug = format!("{stats:?}");
-        assert!(debug.contains("next_sequence_id"));
-    }
-
-    // --- Additional coverage ---
-
-    #[test]
-    fn single_element_send_recv() {
-        let mut ring = SequenceRing::new(1, 1);
-        let sender = ring.sender();
-        sender.send(SequencedRecord::new(1, 42)).unwrap();
-        assert_eq!(ring.recv_ordered().unwrap(), Some(42));
-    }
-
-    #[test]
-    fn capacity_power_of_two_boundary_2() {
-        let mut ring = SequenceRing::new(2, 2);
-        let sender = ring.sender();
-        sender.send(SequencedRecord::new(2, "b")).unwrap();
-        sender.send(SequencedRecord::new(1, "a")).unwrap();
-        assert_eq!(ring.recv_ordered().unwrap(), Some("a"));
-        assert_eq!(ring.recv_ordered().unwrap(), Some("b"));
-    }
-
-    #[test]
-    fn capacity_power_of_two_boundary_4() {
-        let mut ring = SequenceRing::new(4, 4);
-        let sender = ring.sender();
-        for i in (1..=4).rev() {
-            sender.send(SequencedRecord::new(i, i)).unwrap();
-        }
-        for expected in 1..=4u64 {
-            assert_eq!(ring.recv_ordered().unwrap(), Some(expected));
-        }
-    }
-
-    #[test]
-    fn sequence_ids_larger_than_capacity() {
-        let mut ring = SequenceRing::new(4, 4);
-        let sender = ring.sender();
-        // Sequence IDs far beyond capacity
-        sender.send(SequencedRecord::new(1, "a")).unwrap();
-        sender.send(SequencedRecord::new(2, "b")).unwrap();
-        sender.send(SequencedRecord::new(3, "c")).unwrap();
-        sender.send(SequencedRecord::new(4, "d")).unwrap();
-        for expected in ["a", "b", "c", "d"] {
-            assert_eq!(ring.recv_ordered().unwrap(), Some(expected));
-        }
-        // Continue with IDs well past the capacity
-        sender.send(SequencedRecord::new(5, "e")).unwrap();
-        sender.send(SequencedRecord::new(6, "f")).unwrap();
-        assert_eq!(ring.recv_ordered().unwrap(), Some("e"));
-        assert_eq!(ring.recv_ordered().unwrap(), Some("f"));
-        assert_eq!(ring.stats().next_sequence_id, 7);
-    }
-
-    #[test]
-    fn concurrent_senders_from_threads() {
-        use std::thread;
-
-        let mut ring = SequenceRing::new(100, 50);
-        let sender1 = ring.sender();
-        let sender2 = ring.sender();
-
-        let h1 = thread::spawn(move || {
-            for i in (1..=10).step_by(2) {
-                sender1.send(SequencedRecord::new(i, i)).unwrap();
-            }
-        });
-        let h2 = thread::spawn(move || {
-            for i in (2..=10).step_by(2) {
-                sender2.send(SequencedRecord::new(i, i)).unwrap();
-            }
-        });
-
-        h1.join().unwrap();
-        h2.join().unwrap();
-
-        for expected in 1..=10u64 {
-            assert_eq!(ring.recv_ordered().unwrap(), Some(expected));
-        }
-    }
-
-    #[test]
-    fn empty_ring_try_recv_is_empty() {
-        let mut ring = SequenceRing::<i32>::new(8, 4);
-        assert!(matches!(ring.try_recv_ordered(), Err(TryRecvError::Empty)));
-        assert_eq!(ring.stats().next_sequence_id, 1);
-        assert_eq!(ring.stats().reorder_buffer_size, 0);
-    }
-
-    #[test]
-    fn channel_close_drains_buffered_records() {
-        let mut ring = SequenceRing::new(10, 10);
-        let sender = ring.sender();
-        // Send records 3, 2, 1 then close
-        sender.send(SequencedRecord::new(3, "c")).unwrap();
-        sender.send(SequencedRecord::new(2, "b")).unwrap();
-        sender.send(SequencedRecord::new(1, "a")).unwrap();
-        drop(sender);
-        // Internal sender still alive, so recv_ordered works normally
-        assert_eq!(ring.recv_ordered().unwrap(), Some("a"));
-        assert_eq!(ring.recv_ordered().unwrap(), Some("b"));
-        assert_eq!(ring.recv_ordered().unwrap(), Some("c"));
-    }
-
-    #[test]
-    fn try_recv_disconnected_after_drain() {
-        let mut ring = SequenceRing::new(10, 5);
-        let sender = ring.sender();
-        sender.send(SequencedRecord::new(1, "only")).unwrap();
-        drop(sender);
-        // First call returns the record
-        assert_eq!(ring.try_recv_ordered().unwrap(), Some("only"));
-        // After draining, channel still has internal sender so it's Empty not Disconnected
-    }
-
-    #[test]
-    fn interleaved_try_recv_and_send() {
-        let mut ring = SequenceRing::new(10, 5);
-        let sender = ring.sender();
-
-        sender.send(SequencedRecord::new(1, "a")).unwrap();
-        assert_eq!(ring.try_recv_ordered().unwrap(), Some("a"));
-
-        sender.send(SequencedRecord::new(3, "c")).unwrap();
-        assert!(matches!(ring.try_recv_ordered(), Err(TryRecvError::Empty)));
-
-        sender.send(SequencedRecord::new(2, "b")).unwrap();
-        assert_eq!(ring.try_recv_ordered().unwrap(), Some("b"));
-        assert_eq!(ring.try_recv_ordered().unwrap(), Some("c"));
-    }
-
-    #[test]
-    fn sequenced_record_with_complex_data() {
-        let record = SequencedRecord::new(99, vec![vec![1, 2], vec![3, 4]]);
-        assert_eq!(record.sequence_id, 99);
-        assert_eq!(record.data.len(), 2);
-        assert_eq!(record.data[0], vec![1, 2]);
-    }
-
-    #[test]
-    fn stats_after_partial_drain() {
-        let mut ring = SequenceRing::new(10, 10);
-        let sender = ring.sender();
-        sender.send(SequencedRecord::new(1, 1)).unwrap();
-        sender.send(SequencedRecord::new(2, 2)).unwrap();
-        sender.send(SequencedRecord::new(3, 3)).unwrap();
-
-        assert_eq!(ring.recv_ordered().unwrap(), Some(1));
-        let stats = ring.stats();
-        assert_eq!(stats.next_sequence_id, 2);
-    }
-
-    #[test]
-    fn large_window_reorder() {
-        let mut ring = SequenceRing::new(64, 64);
-        let sender = ring.sender();
-
-        // Send 32 records in completely reverse order
-        for i in (1..=32).rev() {
-            sender.send(SequencedRecord::new(i, i)).unwrap();
-        }
-
-        for expected in 1..=32u64 {
-            assert_eq!(ring.recv_ordered().unwrap(), Some(expected));
-        }
-    }
-}
+//! The canonical implementation lives in [`copybook_codec::runtime`]. This
+//! package remains available for consumers migrating from the 0.5 package
+//! layout and contains no sequence-ring implementation of its own.
+
+pub use copybook_codec::runtime::{SequenceRing, SequenceRingStats, SequencedRecord};
