+++ docs/architecture/package-boundary-debt.json
@@ -42,10 +42,6 @@
       "owner_issue": 653
     },
     {
-      "id": "primary-dep:copybook-codec->copybook-sequence-ring",
-      "owner_issue": 654
-    },
-    {
       "id": "primary-dep:copybook-codec->copybook-zoned-format",
       "owner_issue": 653
     },
